Upgrade to released piper version, which had a lot of breaking changes due to shifting modules around and renaming.
