+++ package.json
@@ -14,19 +14,19 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "^4.0.3",
+    "@angular/cdk": "^2.0.0-beta.8",
     "@angular/common": "^4.0.0",
     "@angular/compiler": "^4.0.0",
     "@angular/core": "^4.0.0",
     "@angular/forms": "^4.0.0",
     "@angular/http": "^4.0.0",
     "@angular/material": "^2.0.0-beta.8",
-    "@angular/cdk": "^2.0.0-beta.8",
     "@angular/platform-browser": "^4.0.0",
     "@angular/platform-browser-dynamic": "^4.0.0",
     "@angular/router": "^4.0.0",
     "core-js": "^2.4.1",
     "hammerjs": "^2.0.8",
-    "piper": "github:piper-audio/piper-js",
+    "piper-js": "^0.20.2",
     "recordrtc": "^5.4.2",
     "requirejs": "^2.3.3",
     "rxjs": "^5.4.2",
